Create Mock Title Screen
Created a mock title screen that allows different screens to be loaded. Broke up different screens into different files that return an enum of the next screen when leaving that screen. This system may need to be redone later if we ever need to keep track of data from the last screen or pass data between screens. Added a global header that stores everything needed between all files. Updated the .gitignore.