Center Title Screen
Updated the title screen so that everything is centered and is able to use a custom font which is currently set to Arial. The Play and Options screens will now return to the main menu when escape is pressed. Moved the raylib include into global.h since it will likely be needed for everything that uses global. Updated the makefile to check for changes to global.h when building.